api_image/api_image/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
da-vinci==0.3.0
Django==4.0
django-filter==22.1
django-thumbnails==0.7.0
djangorestframework==3.14.0
drf-nested-serializer==0.1.0
importlib-metadata==6.0.0
Markdown==3.4.1
Pillow==9.4.0
python-decouple==3.7
pytz==2022.7.1
shortuuid==1.0.11
sqlparse==0.4.3
tzdata==2022.7
zipp==3.13.0
